requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
jinja2==3.1.3
playwright==1.41.2
httpx==0.26.0
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3
EOF
